use more usize

+++ src/data_buffers.rs
@@ -10,7 +10,7 @@ use crate::types::{SizedFile, DownloadState, PieceState};
 #[derive(Debug, Eq, PartialEq)]
 pub struct PieceInProgress {
     pub index: u32,
-    pub missing_bytes: u32
+    pub missing_bytes: usize
 }
 
 pub struct Piece {
